Work on email and resource view

+++ thesisfrontend/src/main/java/com/demo/frontend/views/resources/ResourcesView.java
@@ -17,7 +17,9 @@ import com.vaadin.flow.component.dependency.CssImport;
 import com.vaadin.flow.component.dialog.Dialog;
 import com.vaadin.flow.component.grid.Grid;
 import com.vaadin.flow.component.html.Div;
+import com.vaadin.flow.component.html.Span;
 import com.vaadin.flow.component.icon.VaadinIcon;
+import com.vaadin.flow.component.notification.Notification;
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
 import com.vaadin.flow.component.textfield.TextField;
@@ -49,9 +51,9 @@ public class ResourcesView extends VerticalLayout {
 	private AppButton appButton;
 	private QuestionDialog deleteResourceDialog;
 	private List<Resource> resources;
-	private ListDataProvider<Resource> resourceProvider;
 
 	public ResourcesView() {
+		setId("resource-view");
 		setSizeFull();
 		appButton = new AppButton();
 		add(createTopBar());
@@ -65,48 +67,64 @@ public class ResourcesView extends VerticalLayout {
 
 	public void setGrid() {
 		grid = new Grid<>();
+		grid.setId("grid");
 		grid.setSizeFull();
-
 		resources = clientService.getAllResources();
-		resourceProvider = DataProvider.ofCollection(resources);
+		ListDataProvider<Resource> resourceProvider = DataProvider.ofCollection(resources);
 
 		grid.setDataProvider(resourceProvider);
 
 		grid.addColumn(Resource::getName).setHeader("Name").setSortable(true).setKey("Name");
 		grid.addColumn(Resource::getDescription).setHeader("Description");
 		grid.addColumn(Resource::getSeatsAvailable).setHeader("Seats Available");
-
-		if (CurrentUser.isAdmin())
+		grid.addComponentColumn(this::relatedReservation).setHeader("Related Reservation");
+		
+		if (CurrentUser.isAdmin())		
 			grid.addComponentColumn(this::trashIcon);
 
 		add(grid);
+	}
+	
+	public Span relatedReservation(Resource res){
+		int number = clientService.getReservationsByResource(res.getId());
+		
+		Span span = new Span();
+		span.setId("numberRes");
+		span.add(String.valueOf(number));			
+	
+		if(number > 0)
+			span.getElement().getStyle().set("background-color", "#00cc66");
+		
+		return span;
 	}
 
 	public Div trashIcon(Resource res) {
 		Div buttonsContainer = new Div();
 		buttonsContainer.setId("container-icons");
 
-		Button trashButt = appButton.set("", VaadinIcon.TRASH.create());
+		Button trashButt = appButton.set("Delete", VaadinIcon.TRASH.create());
 
 		trashButt.addClickListener(click -> {
-			deleteResourceDialog = new QuestionDialog("Do you want to delete this resource?", VaadinIcon.CHECK.create(),
-					VaadinIcon.CLOSE.create(), "REMOVE");
 
-			/* Delete resource */
-			deleteResourceDialog.getConfirmButton().addClickListener(ev -> {
-				resources.remove(res);
-				HttpEntity<Resource> resource = new HttpEntity<>(res);
-				clientService.deleteResource(resource);
-				grid.getDataProvider().refreshAll();
-				deleteResourceDialog.close();
-			});
-
-			/* Close dialog */
-			deleteResourceDialog.getCancelButton().addClickListener(ev -> {
-				deleteResourceDialog.close();
-			});
+			int number = clientService.getReservationsByResource(res.getId());
+			String text;
+			if (number > 0) {
+				text = "This resource cannot be deleted, there are " + number + " of reservations " + "releted to it";
+				deleteResourceDialog = new QuestionDialog(text, "REMOVE");
+				deleteResourceDialog.getConfirmButton().addClickListener(ev -> deleteResourceDialog.close());
+			} else {
+				text = "Do you want to delete this resource?";
+				deleteResourceDialog = new QuestionDialog(text, "REMOVE");
+				deleteResourceDialog.getConfirmButton().addClickListener(ev -> {
+					resources.remove(res);
+					HttpEntity<Resource> resource = new HttpEntity<>(res);
+					clientService.deleteResource(resource);
+					grid.getDataProvider().refreshAll();
+					deleteResourceDialog.close();
+				});
+			}
 		});
-
+		
 		buttonsContainer.add(trashButt);
 		return buttonsContainer;
 	}
@@ -140,6 +158,7 @@ public class ResourcesView extends VerticalLayout {
 					Resource r = clientService.createResource(resource);
 					resources.add(r);
 					grid.getDataProvider().refreshAll();
+					Notification.show("Resource added");
 					formDialog.close();
 				});
